pyproject.toml
[tool.poetry]
name = "supportbot"
version = "0.1.0"
description = ""
authors = ["Oleg Kuzmenko <olegkuzmenko2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
aiogram = "2.11.2"
sqlalchemy = "^2.0.23"
greenlet = "^3.0.1"
telethon = "^1.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
